% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{default.tables}
\alias{default.tables}
\title{Model input tables}
\format{
A list with 6 elements:
\describe{
   \item{\code{post.sbw.reg}: A data frame of species regeneration probability after mortality by SBW defoliation
     \itemize{
        \item{\code{spp}: Code of the species.}
        \item{\code{potential.spp}: Code of the species to change.}
        \item{\code{ptrans}: Probability of transition.}
      }
   }
   \item{\code{forest.succ}: A data frame of transition probability of species seral succession
     \itemize{
        \item{\code{spp}: Code of the species.}
        \item{\code{potential.spp}: Code of the species to change.}
        \item{\code{ptrans}: Probability of transition.}
      }
   }  
   \item{\code{spp.colonize.persist}: A data frame of species seral succession
     \itemize{
        \item{\code{spp}: Code of the species.}
        \item{\code{rad}: Estimated maximum colonization distance (in m).}
        \item{\code{nneigh}: Minimum number of source cells within the colonization distance to enable colonization.}
        \item{\code{persist}:It indicates whether we allow the transition probability to remain high locally.}
      }
   }  
   \item{\code{temp.suitability}: A data frame with thresholds of temperature to define optimal and good species niche
     \itemize{
        \item{\code{spp}: Code of the species.}
        \item{\code{th1}: Temperature lower threshold to define 'good' species niche (in ºC).}
        \item{\code{th2}: Temperature lower threshold to define 'optimal' species niche (in ºC).}
        \item{\code{th3}: Temperature upper threshold to define 'good' species niche (in ºC).}
        \item{\code{th4}: Temperature upper threshold to define 'optimal' species niche (in ºC).}
      }
   }  
   \item{\code{prec.suitability}: A data frame with thresholds of precipitation to define optimal and good species niche
     \itemize{
        \item{\code{spp}: Code of the species.}
        \item{\code{th1}: Precipitation lower threshold to define 'good' species niche (in mm).}
        \item{\code{th2}: Precipitation lower threshold to define 'optimal' species niche(in mm).}
        \item{\code{th3}: Precipitation upper threshold to define 'good' species niche (in mm).}
        \item{\code{th4}: Precipitation upper threshold to define 'optimal' species niche (in mm).}
      }
   }  
   \item{\code{soil.suitability}: A data frame with thresholds of precipitation to define optimal and good species niche
     \itemize{
        \item{\code{spp}: Code of the species.}
        \item{\code{thT}: Species suitability in glacial tills [0,1].}
        \item{\code{thO}: Species suitability in organic deposists [0,1].}
        \item{\code{thR}: Species suitability in rock outcrops or very thin deposits [0,1].}
        \item{\code{thS}: Species suitability in sandy deposists [0,1].}
        \item{\code{thA}: Species suitability in clay deposits [0,1].}
      }
   }  
 }
}
\usage{
default.tables
}
\description{
List of data frames being input tables of the SBW model
}
\keyword{datasets}
