% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate causal contrasts and population means}
\usage{
estimate(object, digits = 6, ...)
}
\arguments{
\item{object}{an object from function \code{\link[sbw]{sbw}}.}

\item{digits}{a scalar with the number of significant digits used to display the estimates. The default is 6.}

\item{...}{ignored arguments}
}
\description{
Function for estimating causal contrasts and population means using the output from \code{\link[sbw]{sbw}}.
}
\examples{
# Please see the examples in sbw.
}
