% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.mortality.r
\name{forest.mortality}
\alias{forest.mortality}
\title{Forest mortality by SBW}
\usage{
forest.mortality(land)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}
}
\value{
A vector with the \code{cell.id} of the forest cells killed by 
recurrent SBW defoliation according to species composition
}
\description{
Determines the cells killed by recurrent SBW defoliation
}
\examples{
data(landscape)
forest.mortality(landscape)

}
