% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Visualize output from \code{sbw}}
\usage{
summarize(object, digits = 6, ...)
}
\arguments{
\item{object}{an object from the class \code{sbwcau} or \code{sbwpop} obtained after using \code{\link[sbw]{sbw}}.}

\item{digits}{The number of significant digits that will be displayed. The default is 6.}

\item{...}{ignored arguments}
}
\description{
Function for summarizing the output from \code{\link[sbw]{sbw}}.
}
\examples{
# Please see the examples in function sbw.
}
