% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize output from \code{sbw}}
\usage{
visualize(object, plot_cov, ...)
}
\arguments{
\item{object}{an object from function \code{\link[sbw]{sbw}}.}

\item{plot_cov}{names of covariates for which balance is to be displayed.  If \code{NULL}, all of the covariates will be displayed.}

\item{...}{ignored}
}
\description{
Function for visualizing the output from \code{\link[sbw]{sbw}}.
}
\examples{
# Please see the examples in \\code{sbw}.
}
