% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBW.r
\docType{package}
\name{SBW}
\alias{SBW}
\title{SBW: A package for simulating spruce budworm outbreaks in Quebec province}
\description{
The SBW package provides a set of functions to simulate four different phases of 
spruce budworm outbreaks: pre-epidemic, epidemic, collapse, and calm, as well as the 
landscape-scale processes associated to these (e.g. neighbour contagion, long-term dispersal).
It also includes the effects on forest communities (i.e. mortality induced by SBW recurrent defoliation).
}
\section{Author(s)}{

\bold{Maintainer}: Núria Aquilué \email{nuria.aquilue@ctfc.cat}  

\bold{Authors}: Núria Aquilué, Mathieu Bouchard and Élise Filotas
}

